Fixed gameloop correctly, spread classes out into proper file structure
